refactor background/methods folder structure
