+++ src/background/methods/dataTypes/arrays.ts
