+++ src/background/methods/dataTypes/bytes.ts
