+++ src/background/methods/dataTypes/objects/objectEquals.ts
