+++ src/background/methods/dataTypes/strings.ts
