+++ src/background/methods/dataTypes/time.ts
