+++ src/components/pages/filesytem/FileListItem.tsx
@@ -2,8 +2,8 @@ import {FsEntity} from "../../../background/api/filesystemTypes";
 import React, {ReactElement} from "react";
 import {Col, Form} from "react-bootstrap";
 import {Link} from "react-router-dom";
-import {getDateAsStringFromTimestamp} from "../../../background/methods/time";
-import {formatBytes} from "../../../background/methods/bytes";
+import {getDateAsStringFromTimestamp} from "../../../background/methods/dataTypes/time";
+import {formatBytes} from "../../../background/methods/dataTypes/bytes";
 import {connect, ConnectedProps} from "react-redux";
 import {SystemState} from "../../../background/redux/actions/sytemState";
 import {addToSelected, removeFromSelected} from "../../../background/redux/actions/filesystem";
