+++ src/components/pages/filesytem/upload/UploadZone.tsx
@@ -6,7 +6,7 @@ import { Modal } from "react-bootstrap";
 import { uploadFiles, uploadPreflight } from "../../../../background/api/filesystem";
 import { FsEntity } from "../../../../background/api/filesystemTypes";
 import { UploadDecisionsModalContent } from "./UploadDecisionsModalContent";
-import { divideArrayByCondition } from "../../../../background/methods/arrays";
+import { divideArrayByCondition } from "../../../../background/methods/dataTypes/arrays";
 import { getPathWithoutName, isFileNameValid } from "../../../../background/methods/filesystem";
 import {
   EditableEntityError,
